requirements.txt
yfinance>=0.2.38
pandas>=2.1.0
numpy>=1.24.0
scipy>=1.11.0
python-telegram-bot>=21.0
APScheduler>=3.10.0
SQLAlchemy>=2.0.0
openpyxl>=3.1.0
python-dotenv>=1.0.0
ta>=0.11.0
